refactor(RaceGame) : 전체적인 클래스 구조 수정
- RaceController에 main 함수 구현
- RaceGame의 race 메서드가 시도 횟수와 자동차 이름 목록을 매개변수로 받도록 구현

(cherry picked from commit b423d8977f41edb50680177b3ff35e8e9936ebad)
(cherry picked from commit 9b38f4107d95fd04db416e11b409016d7dd89d18)
