--- src/main/java/Application.java
@@ -1,9 +1,0 @@
-import contoller.RaceController;
-
-public class Application {
-
-  public static void main(String[] args) {
-    RaceController raceController = new RaceController();
-    raceController.run();
-  }
-}
